+++ main.py
@@ -7,16 +7,31 @@ This is:
 """
 import logging
 import signal
+import random
 
-from pyhap.accessory import Bridge
+from pyhap.accessory import Accessory, Bridge
 from pyhap.accessory_driver import AccessoryDriver
 import pyhap.loader as loader
 from pyhap import camera
+from pyhap.const import CATEGORY_SENSOR
 
-# The below package can be found in the HAP-python github repo under accessories/
-from accessories.TemperatureSensor import TemperatureSensor
+logging.basicConfig(level=logging.INFO, format="[%(module)s] %(message)s")
 
-logging.basicConfig(level=logging.INFO)
+
+class TemperatureSensor(Accessory):
+    """Fake Temperature sensor, measuring every 3 seconds."""
+
+    category = CATEGORY_SENSOR
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        serv_temp = self.add_preload_service('TemperatureSensor')
+        self.char_temp = serv_temp.configure_char('CurrentTemperature')
+
+    @Accessory.run_at_interval(3)
+    async def run(self):
+        self.char_temp.set_value(random.randint(18, 26))
 
 
 def get_bridge(driver):